modify signup form
