+++ web/src/components/Auth/Common/index.tsx
@@ -1,0 +1,2 @@
+export * from "./BlankLayoutWrapper.tsx";
+export * from "./Card.tsx";
